requirements-cuda.txt
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
transformers
peft
pytesseract
Pillow
pdfminer.six
langdetect
googletrans==4.0.0-rc1
quart
quart-cors
streamlit